+++ packages/shared/src/index.ts
@@ -6,5 +6,5 @@ export * from './assets-transpilers';
 export * from './utils';
 export * from './module-resolver';
 export * from './common';
-export * from './error';
+export * from './reporter';
 export * from './fetch-utils/wrapFetchWithCache';
